requirements.txt
streamlit==1.44.1
pandas==19.0.1
google-generativeai==0.8.4
python-dotenv==1.1.0
markdown==3.6
